Bug fix : unappropriate data type used for LREAL coding
git-svn-id: svn://svn.code.sf.net/p/enipexplorer/code@41 450f7cbe-2bfd-44d8-9b4c-35985d01ad46

+++ EnIPStack/ObjectsLibrary/CIPObject.cs
@@ -217,7 +217,7 @@ namespace System.Net.EnIPStack.ObjectsLibrary
             Idx += 8;
             return ret;
         }
-        public void SetDouble(ref int Idx, byte[] buf, UInt64? val)
+        public void SetDouble(ref int Idx, byte[] buf, Double? val)
         {
             if (val == null) return;
             Array.Copy(BitConverter.GetBytes(val.Value), 0, buf, Idx, 8);
